added authorization to delete function

+++ backend/routes/posts.js
@@ -100,6 +100,7 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, 
   })
   Post.updateOne({_id : req.params.id, creator: req.userData.userId }, post)
     .then(result =>{
+      //uses nModified -- print console.log(result) if unclear
       if(result.nModified > 0){
         res.status(200).json({
           message: "successful update",
@@ -113,12 +114,18 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, 
 });
 
 router.delete('/:id',checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
     .then(result =>{
-      console.log(result);
-      res.status(200).json({
-        message: "Deleted"
-      });
+    //delete field uses n as it doesn't have nModified-- print console.log(result) if unclear
+      if(result.n > 0){
+        res.status(200).json({
+          message: "Deleted",
+        })
+      } else{
+        res.status(401).json({
+          message: "The delete is not successful",
+        })
+      }
     })
 });
